feat: added common operatorFilter tests
